# openapi.yaml
openapi: 3.1.0
info:
  title: Incident Tracker API
  version: 0.3.0
  description: >-
    Multi-tenant incident tracker API with session auth, tenancy, and incident workflows.
    ID-based routes return 404 for objects that are missing or inaccessible in the active
    tenant context, and return 403 only when object existence is established but permission
    is denied.
servers:
  - url: http://localhost:3000
paths:
...
  /v1/auth/login:
    post:
      summary: Login with email and password
      operationId: login
      security:
        - CsrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthStateAuthenticatedResponse'
        '401':
          description: Invalid credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: CSRF validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '423':
          description: Account temporarily locked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/auth/logout:
    post:
      summary: Logout and clear session
      operationId: logout
      security:
        - SessionCookie: []
          CsrfToken: []
      responses:
        '204':
          description: Logged out.
        '403':
          description: CSRF validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/auth/password/forgot:
    post:
      summary: Request password reset
      operationId: forgotPassword
      security:
        - CsrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '202':
          description: Password reset accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForgotPasswordResponse'
        '403':
          description: CSRF validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/auth/password/reset:
    post:
      summary: Reset password using token
      operationId: resetPassword
      security:
        - CsrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '204':
          description: Password reset successful.
        '400':
          description: Invalid reset token or weak password.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: CSRF validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/tenants/{tenantId}/service-accounts:
    get:
      summary: List service accounts for a tenant
      operationId: listServiceAccounts
      security:
        - SessionCookie: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200':
          description: Service accounts.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceAccountListResponse'
        '403':
          description: Permission denied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Tenant not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create a service account
      operationId: createServiceAccount
      security:
        - SessionCookie: []
          CsrfToken: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateServiceAccountRequest'
      responses:
        '201':
          description: Service account created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceAccountItemResponse'
        '403':
          description: Permission denied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Tenant not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/tenants/{tenantId}/api-keys:
    get:
      summary: List API keys for a tenant
      operationId: listApiKeys
      security:
        - SessionCookie: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200':
          description: API keys.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyListResponse'
        '403':
          description: Permission denied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Tenant not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create an API key
      operationId: createApiKey
      security:
        - SessionCookie: []
          CsrfToken: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApiKeyRequest'
      responses:
        '201':
          description: API key created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateApiKeyResponse'
        '403':
          description: Permission denied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Tenant or service account not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/tenants/{tenantId}/api-keys/{apiKeyId}/revoke:
    post:
      summary: Revoke an API key
      operationId: revokeApiKey
      security:
        - SessionCookie: []
          CsrfToken: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/ApiKeyId'
      responses:
        '200':
          description: API key revoked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyItemResponse'
        '403':
          description: Permission denied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: API key not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  parameters:
    TenantId:
      name: tenantId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    ApiKeyId:
      name: apiKeyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: incident_tracker_sid
    CsrfToken:
      type: apiKey
      in: header
      name: x-csrf-token
    ApiKeyHeader:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    ForgotPasswordResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          const: AUTH_PASSWORD_RESET_REQUESTED
        message:
          type: string
        resetToken:
          type: string
    ResetPasswordRequest:
      type: object
      required: [token, newPassword]
      properties:
        token:
          type: string
          minLength: 16
        newPassword:
          type: string
          minLength: 12
...
    AuthType:
      type: string
      enum: [session, api_key]
    ApiKeyScope:
      type: string
      enum: [read, write]
...
    AuthenticatedUser:
      type: object
      required: [id, email, passwordUpdatedAt, createdAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        passwordUpdatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    AuthStateAuthenticatedResponse:
      type: object
      required: [authenticated, authType, user, csrfToken]
      properties:
        authenticated:
          type: boolean
          const: true
        authType:
          $ref: '#/components/schemas/AuthType'
        user:
          $ref: '#/components/schemas/AuthenticatedUser'
        csrfToken:
          type: string
...
    ServiceAccount:
      type: object
      required: [id, tenantId, name, ownerUserId, createdByUserId, revokedAt, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        name:
          type: string
        ownerUserId:
          type: string
          format: uuid
        createdByUserId:
          type: string
          format: uuid
        revokedAt:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ApiKeyRecord:
      type: object
      required:
        [id, tenantId, serviceAccountId, name, keyPrefix, scopes, createdByUserId, lastUsedAt, revokedAt, createdAt]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        serviceAccountId:
          type: string
          format: uuid
        name:
          type: string
        keyPrefix:
          type: string
        scopes:
          type: array
          items:
            $ref: '#/components/schemas/ApiKeyScope'
        createdByUserId:
          type: string
          format: uuid
        lastUsedAt:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
        revokedAt:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
        createdAt:
          type: string
          format: date-time
    CreateServiceAccountRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 120
        ownerUserId:
          type: string
          format: uuid
    ServiceAccountListResponse:
      type: object
      required: [serviceAccounts, csrfToken]
      properties:
        serviceAccounts:
          type: array
          items:
            $ref: '#/components/schemas/ServiceAccount'
        csrfToken:
          type: string
    ServiceAccountItemResponse:
      type: object
      required: [serviceAccount, csrfToken]
      properties:
        serviceAccount:
          $ref: '#/components/schemas/ServiceAccount'
        csrfToken:
          type: string
    CreateApiKeyRequest:
      type: object
      required: [serviceAccountId, name, scopes]
      properties:
        serviceAccountId:
          type: string
          format: uuid
        name:
          type: string
          minLength: 3
          maxLength: 120
        scopes:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/ApiKeyScope'
    ApiKeyListResponse:
      type: object
      required: [apiKeys, csrfToken]
      properties:
        apiKeys:
          type: array
          items:
            $ref: '#/components/schemas/ApiKeyRecord'
        csrfToken:
          type: string
    ApiKeyItemResponse:
      type: object
      required: [apiKey, csrfToken]
      properties:
        apiKey:
          $ref: '#/components/schemas/ApiKeyRecord'
        csrfToken:
          type: string
    CreateApiKeyResponse:
      type: object
      required: [apiKey, secret, redactedSecret, csrfToken]
      properties:
        apiKey:
          $ref: '#/components/schemas/ApiKeyRecord'
        secret:
          type: string
        redactedSecret:
          type: string
        csrfToken:
          type: string
...
    ErrorResponse:
      type: object
      required: [code, message, traceId]
      properties:
        code:
          type: string
        message:
          type: string
        traceId:
          type: string
        details:
          description: Optional structured metadata for debugging and validation.